Refactor : MemberId, UpdateRequest 변수 수정  (#68)
