+++ src/main/java/org/sesac/slopedbe/member/repository/MemberRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 	Optional<Member> findByEmail(String email);
-	Optional<Member> findByMemberId(String id);
+	Optional<Member> findByMemberId(String memberId);
 	void deleteByEmail(String email);
 
-	boolean existsByMemberId(String id);
+	boolean existsByMemberId(String memberId);
 
 	boolean existsByEmail(String email);
 }
